Week3 Day2

+++ hotelapp/src/main/java/com/visa/prj/service/BookingService.java
@@ -15,7 +15,7 @@ public class BookingService {
 	@Autowired
 	private BookingDao bookingDao;
 
-	public Hotel getHotelById(int id) {
+	public Hotel getHotelById(long id) {
 		return bookingDao.findHotelById(id);
 	}
